MXS-3233 Move config class to global scope
Will be needed elsewhere as well, so better not to have it nested
inside one particular class.
