+++ server/modules/protocol/MongoDB/config.hh
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2020 MariaDB Corporation Ab
+ *
+ * Use of this software is governed by the Business Source License included
+ * in the LICENSE.TXT file and at www.mariadb.com/bsl11.
+ *
+ * Change Date: 2024-10-14
+ *
+ * On the date above, in accordance with the Business Source License, use
+ * of this software will be governed by version 2 or later of the General
+ * Public License.
+ */
+#pragma once
+
+#include "mongodbclient.hh"
+#include <maxscale/config2.hh>
+
+class Config : public mxs::config::Configuration
+{
+public:
+    Config();
+    Config(Config&&) = default;
+
+    std::string user;
+    std::string password;
+
+    static mxs::config::Specification& specification();
+};
