+++ server/modules/protocol/MongoDB/protocolmodule.hh
@@ -14,21 +14,12 @@
 
 #include "mongodbclient.hh"
 #include <maxscale/protocol2.hh>
-#include <maxscale/config2.hh>
+#include "config.hh"
+
 
 class ProtocolModule : public mxs::ProtocolModule
 {
 public:
-    class Config : public mxs::config::Configuration
-    {
-    public:
-        Config();
-        Config(Config&&) = default;
-
-        std::string user;
-        std::string password;
-    };
-
     using AuthenticatorList = std::vector<mxs::SAuthenticatorModule>;
 
     static ProtocolModule* create(const mxs::ConfigParameters& params);
@@ -48,8 +39,6 @@ public:
     std::unique_ptr<mxs::UserAccountManager> create_user_data_manager() override;
 
     AuthenticatorList create_authenticators(const mxs::ConfigParameters& params) override;
-
-    static mxs::config::Specification specification;
 
 private:
     ProtocolModule(Config&& config);
